Changed I2C to use Zynq PS I2C controller
The design now recognises the devices on the I2C bus by using the
inbuilt Zynq I2C, but hdmi is still not coming up.  The devices on the
AXI bus are currently not responding (which is why the PL I2C bus isn't
responding).  Because of the change, the device tree has been added.
I've also added a u-boot and Linux image.  I didn't build these, they
are copied directly from the adi quickstart image.  I will update these
after the system is fully working.